Move remaining views into subfolder
